# docs/openapi.yaml
---
openapi: 3.0.0
info:
  title: AR.IO ArNS Resolver
  version: 0.0.1
  description: |
    AR.IO ArNS Resolver
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: apiToken
      description: ADMIN_API_KEY set in your .env file.
  schemas:
    ArweaveAddress:
      type: string
    ArweaveWallet:
      type: string
      nullable: true
    ArweaveId:
      type: string
      nullable: true
    Timestamp:
      type: integer
      nullable: true
    Evaluation:
      type: boolean
    HealthCheck:
      type: object
      properties:
        uptime: { type: number }
        date: { type: string, format: date }
        message: { type: string }
    URL:
      type: string
      description: |
        A URL to a resource
    Info:
      type: object
      properties:
        processId: { '$ref': '#/components/schemas/ArweaveId' }
        lastEvaluationTimestamp: { '$ref': '#/components/schemas/Timestamp' }
    Record:
      type: object
      properties:
        name: { type: string }
        ttlSeconds: { type: integer }
        txId: { $ref: '#/components/schemas/ArweaveId' }
        endTimestamp: { '$ref': '#/components/schemas/Timestamp' }
        owner: { $ref: '#/components/schemas/ArweaveAddress' }
        processId: { $ref: '#/components/schemas/ArweaveId' }
        type: { type: string }

paths:
  '/ar-io/resolver/healthcheck':
    get:
      responses:
        '200':
          description: |-
            200 response
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/HealthCheck'
  '/ar-io/resolver/info':
    get:
      responses:
        '200':
          description: |-
            200 response
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/Info'
  '/ar-io/resolver/admin/evaluate':
    post:
      security:
        - bearerAuth: []
      responses:
        '200':
          description: |-
            200 response
          content:
            application/json:
              schema:
                    type: object
                    properties:
                      message: { type: string }
                      processId: { $ref: '#/components/schemas/ArweaveId' }
        '202':
          description: |-
            202 response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        '400':
          description: |-
            400 response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        '401':
          description: |-
            401 response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        '500':
          description: |-
            500 response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
  '/ar-io/resolver/records/{name}':
    head:
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: |-
            The name of the record to resolve
      responses:
        '200':
          description: |-
            200 response
        '404':
          description: |-
            404 response
        '500':
          description: |-
            500 response
    get:
      responses:
        '200':
          description: |-
            200 response
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/Record'
        '404':
          description: |-
            404 response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        '500':
          description: |-
            500 response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }